+++ front/src/pages/lege-oharra.tsx
@@ -1,30 +1,19 @@
 import React from 'react';
 
-import { graphql, PageProps, useStaticQuery } from 'gatsby';
+import { PageProps } from 'gatsby';
 
 import { SEO } from '../components/SEO';
+import { useLegeOharraContent } from '../viewQueries/useLegeOharraContent';
 import { RegularPage } from '../views/RegularPage';
-import { RegularPageContent } from '../views/RegularPage/RegularPageContent';
-
-interface DataProps {
-  strapiLegeOharra: RegularPageContent;
-}
 
 const LegeOharra: React.FC<PageProps> = ({ location }) => {
-  const { strapiLegeOharra } = useStaticQuery<DataProps>(graphql`
-    {
-      strapiLegeOharra {
-        izenburua
-        edukia
-      }
-    }
-  `);
+  const content = useLegeOharraContent();
 
   return (
     <>
-      <SEO title={strapiLegeOharra.izenburua} location={location} />
+      <SEO title={content.izenburua} location={location} />
 
-      <RegularPage content={strapiLegeOharra} />
+      <RegularPage content={content} />
     </>
   );
 };
