+++ front/src/viewQueries/useLegeOharraContent.tsx
@@ -1,0 +1,20 @@
+import { graphql, useStaticQuery } from 'gatsby';
+
+import { RegularPageContent } from '../views/RegularPage/RegularPageContent';
+
+interface DataProps {
+  strapiLegeOharra: RegularPageContent;
+}
+
+export function useLegeOharraContent(): RegularPageContent {
+  const { strapiLegeOharra } = useStaticQuery<DataProps>(graphql`
+    {
+      strapiLegeOharra {
+        izenburua
+        edukia
+      }
+    }
+  `);
+
+  return strapiLegeOharra;
+}
